+++ SMASHING/Presentation/Profile/ProfileViewController.swift
@@ -20,12 +20,3 @@ final class ProfileViewController: BaseViewController {
     }
 
 }
-
-// MARK: - Preview
-
-import SwiftUI
-
-@available(iOS 18.0, *)
-#Preview {
-    ProfileViewController()
-}
